+++ src/Components/offline status dialog/docs/changelog.md
@@ -1,5 +1,4 @@
 # 9th July 2026
-
 - Built offline status dialog
 - Rendered info dynamically
 - Built timer
@@ -7,3 +6,13 @@
 - Designed to adapt to change of state
 - Made dynamic classes
 - Built continously running function
+
+# 10th July 2026
+- Fixed timer issue
+- Fixed the whole component
+- Made button to run function
+- Created offline and online state
+- Makes use of endpoint to be called once
+- Made function dynamic by use of parameters
+- Consists of three functions working together
+- Determines what to display based on internet status